# Phase 3 & 4: I/O & Accelerator Fast Paths + Build Optimization & AI Tuning