--- a/moduler_kernel/performance/phase3/gpu/gpu_device.c
+++ b/moduler_kernel/performance/phase3/gpu/gpu_device.c
@@ -1,0 +1,15 @@
+#include "gpu_device.h"
+#include <stdlib.h>
+
+int gpu_init(void) {
+    return 0;
+}
+
+void gpu_shutdown(void) {
+}
+
+int gpu_probe_devices(gpu_device_t** devices, size_t max_devices) {
+    (void)devices;
+    (void)max_devices;
+    return 0;
+}