--- a/moduler_kernel/performance/phase3/gpu/gpu_kernel.c
+++ b/moduler_kernel/performance/phase3/gpu/gpu_kernel.c
@@ -1,0 +1,4 @@
+#include "gpu_kernel.h"
+
+int gpu_kernel_init(void) { return 0; }
+void gpu_kernel_shutdown(void) {}