--- a/moduler_kernel/performance/phase3/gpu/gpu_memory.c
+++ b/moduler_kernel/performance/phase3/gpu/gpu_memory.c
@@ -1,0 +1,4 @@
+#include "gpu_memory.h"
+
+int gpu_memory_init(void) { return 0; }
+void gpu_memory_shutdown(void) {}