--- a/moduler_kernel/performance/phase3/gpu/gpu_memory.h
+++ b/moduler_kernel/performance/phase3/gpu/gpu_memory.h
@@ -1,0 +1,17 @@
+#ifndef PHASE3_GPU_MEMORY_H
+#define PHASE3_GPU_MEMORY_H
+
+#include <stdint.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+int gpu_memory_init(void);
+void gpu_memory_shutdown(void);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif