--- a/moduler_kernel/performance/phase3/net/mbuf.c
+++ b/moduler_kernel/performance/phase3/net/mbuf.c
@@ -1,0 +1,4 @@
+#include "mbuf.h"
+
+int mbuf_pool_init(void) { return 0; }
+void mbuf_pool_shutdown(void) {}