--- a/moduler_kernel/performance/phase3/net/mbuf.h
+++ b/moduler_kernel/performance/phase3/net/mbuf.h
@@ -1,0 +1,19 @@
+#ifndef PHASE3_MBUF_H
+#define PHASE3_MBUF_H
+
+#include <stdint.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+typedef struct mbuf mbuf_t;
+
+int mbuf_pool_init(void);
+void mbuf_pool_shutdown(void);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif