--- a/moduler_kernel/performance/phase3/net/net_device.c
+++ b/moduler_kernel/performance/phase3/net/net_device.c
@@ -1,0 +1,15 @@
+#include "net_device.h"
+#include <stdlib.h>
+
+int net_init(void) {
+    return 0;
+}
+
+void net_shutdown(void) {
+}
+
+int net_probe_devices(net_device_t** devices, size_t max_devices) {
+    (void)devices;
+    (void)max_devices;
+    return 0;
+}