--- a/moduler_kernel/performance/phase3/net/pmd.c
+++ b/moduler_kernel/performance/phase3/net/pmd.c
@@ -1,0 +1,4 @@
+#include "pmd.h"
+
+int pmd_init(void) { return 0; }
+void pmd_shutdown(void) {}