… moduler_kernel/performance/PHASE3_4_INTEGRATION.md
… moduler_kernel/performance/phase3/CMakeLists.txt
… moduler_kernel/performance/phase3/README.md
… moduler_kernel/performance/phase3/bench/bench_phase3.c
… moduler_kernel/performance/phase3/docs/API.md
… moduler_kernel/performance/phase3/docs/ARCHITECTURE.md
… moduler_kernel/performance/phase3/nvme/nvme_io.h
… moduler_kernel/performance/phase3/nvme/nvme_queue.c
… moduler_kernel/performance/phase3/nvme/nvme_queue.h
… moduler_kernel/performance/phase3/tests/test_nvme_queue.c
… moduler_kernel/performance/phase4/CMakeLists.txt
… moduler_kernel/performance/phase4/README.md
… moduler_kernel/performance/phase4/bench/bench_phase4.c
… moduler_kernel/performance/phase4/docs/ARCHITECTURE.md
… moduler_kernel/performance/phase4/docs/BUILD.md
… moduler_kernel/performance/phase4/intrinsics/cpu_features.c
… moduler_kernel/performance/phase4/pgo/pgo_profile.h
… moduler_kernel/performance/phase4/profiler/ai_optimizer.c
… moduler_kernel/performance/phase4/profiler/ai_optimizer.h
… moduler_kernel/performance/phase4/profiler/perf_collector.c
… moduler_kernel/performance/phase4/profiler/perf_collector.h
… moduler_kernel/performance/phase4/tests/test_cpu_features.c
… moduler_kernel/performance/phase4/tests/test_perf_collector.c